python-common,mgr/cephadm: move assert_valid_host to service_spec
