+++ src/python-common/ceph/deployment/service_spec.py
@@ -16,6 +16,18 @@ class ServiceSpecValidationError(Exception):
         super(ServiceSpecValidationError, self).__init__(msg)
 
 
+def assert_valid_host(name):
+    p = re.compile('^[a-zA-Z0-9-]+$')
+    try:
+        assert len(name) <= 250, 'name is too long (max 250 chars)'
+        for part in name.split('.'):
+            assert len(part) > 0, '.-delimited name component must not be empty'
+            assert len(part) <= 63, '.-delimited name component must not be more than 63 chars'
+            assert p.match(part), 'name component must include only a-z, 0-9, and -'
+    except AssertionError as e:
+        raise ServiceSpecValidationError(e)
+
+
 class HostPlacementSpec(namedtuple('HostPlacementSpec', ['hostname', 'network', 'name'])):
     def __str__(self):
         res = ''
@@ -99,8 +111,11 @@ class HostPlacementSpec(namedtuple('HostPlacementSpec', ['hostname', 'network', 
             except ValueError as e:
                 # logging?
                 raise e
-
+        host_spec.validate()
         return host_spec
+
+    def validate(self):
+        assert_valid_host(self.hostname)
 
 
 class PlacementSpec(object):
@@ -198,6 +213,8 @@ class PlacementSpec(object):
             raise ServiceSpecValidationError("num/count must be > 1")
         if self.host_pattern and self.hosts:
             raise ServiceSpecValidationError('cannot combine host patterns and hosts')
+        for h in self.hosts:
+            h.validate()
 
     @classmethod
     def from_string(cls, arg):
